modulo_dos/smart-mobility-hub-old/smart-mobility-hub/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
pytest==8.3.3
httpx==0.27.2